+++ src/logger.ts
@@ -15,7 +15,7 @@ import {
 } from './constants';
 import { readYML, isEmpty } from './utils/common';
 import { generateSettings, LOGS_METADATA, PLUGIN_USER } from "./settings";
-import { UserUtils } from './utils/user_utils';
+import { UserState } from './utils/user_utils';
 
 let macAddress = "";
 macaddress.one().then(mac => macAddress = mac);
@@ -86,8 +86,8 @@ const putLogEvent = async (msg: string, eventType: string, additionalMsg="", log
 			secretKey = user.secret_key;
 		}
 	} else {
-		const userUtils = new UserUtils();
-		const activeUser = userUtils.getActiveUser();
+		const userState = new UserState();
+		const activeUser = userState.getUser();
 		if (activeUser) {
 			email = activeUser.email;
 			accessKey = users[email].access_key;