+++ src/utils/user_utils.ts
@@ -1,7 +1,57 @@
 import fs from 'fs';
 import { generateSettings } from "../settings";
 import { isUserActive, readYML } from './common';
-import { IUser } from '../interface';
+import { IUser, IUserState } from '../interface';
+import { CODESYNC_STATES, CodeSyncState } from './state_utils';
+
+
+export class UserState {
+
+	set = (isUserActive: boolean, isDeactivated: boolean): void => {
+		CodeSyncState.set(CODESYNC_STATES.IS_USER_ACTIVE, isUserActive);
+		CodeSyncState.set(CODESYNC_STATES.ACCOUNT_DEACTIVATED, isDeactivated);
+	}
+
+	get = (): IUserState => {
+		const isActive = CodeSyncState.get(CODESYNC_STATES.IS_USER_ACTIVE);
+		const isDeactivated = CodeSyncState.get(CODESYNC_STATES.ACCOUNT_DEACTIVATED);
+		return {
+			isActive,
+			isDeactivated
+		};
+	}
+
+	setDeactivated = (): void => {
+		const isActive = true;
+		const isDeactivated = true;
+		this.set(isActive, isDeactivated);
+	}
+
+	setValidAccount = (): void => {
+		const isActive = true;
+		const isDeactivated = false;
+		this.set(isActive, isDeactivated);
+	}
+
+	isValidAccount = (): boolean => {
+		const state = this.get();
+		return state.isActive && !state.isDeactivated;
+	}
+
+	isDeactivated = (): boolean => {
+		const state = this.get();
+		return state.isActive && state.isDeactivated;
+	}
+
+	getUser = (checkState=true): IUser|null => {
+		if (checkState) {
+			const state = this.get();
+			if (!state.isActive) return null;	
+		}
+		const userUtils = new UserUtils();
+		return userUtils.getActiveUser();
+	}
+}
 
 export class UserUtils {
 	users = <any>{};